backend/ml_service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
numpy==1.26.4
scipy==1.11.4
pandas==2.2.0
pyarrow==15.0.0
implicit==0.7.2
sentence-transformers==2.5.1
faiss-cpu==1.7.4
torch==2.2.1
lightgbm==4.3.0
scikit-learn==1.4.1.post1
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
redis==5.0.1
httpx==0.27.0
prometheus-client==0.20.0
loguru==0.7.2
pytest==8.0.1
pytest-asyncio==0.23.5
boto3==1.34.51
